moved allow duplicates from ic to all
