+++ dataquality/loggers/logger_config/base_logger_config.py
@@ -36,7 +36,11 @@ class BaseLoggerConfig(BaseModel):
     finish: Callable = lambda: None  # Overwritten in Semantic Segmentation
     # True when calling `init` with a run that already exists
     existing_run: bool = False
-    dataloader_random_sampling = False
+    dataloader_random_sampling = True
+    # Keep track of the ids that have been observed in the current epoch
+    # the key is the split and epoch like observed_ids["train_0"] = {0, 1, 2, 3}
+    observed_ids: Dict[str, set] = dict()
+    all_ids: Dict[str, set] = dict()
 
     class Config:
         validate_assignment = True
